# lesson12/ClinicService/ClinicService/Models/Responses/ClientResponse.cs
// Client/Responses/ClientResponse.cs
namespace ClinicService.Models.Responses
{
    public class ClientResponse
    {
        public int ClientId { get; set; }
        public string Document { get; set; }
        public string SurName { get; set; }
        public string FirstName { get; set; }
        public string? Patronymic { get; set; }
        public DateTime Birthday { get; set; }
    }
}