# lesson12/ClinicService/ClinicService/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ClinicService API",
    "description": "API для управления ветеринарной клиникой",
    "version": "1.0.0",
    "contact": {
      "name": "Support",
      "email": "support@clinic.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5076",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/Client/create": {
      "post": {
        "tags": [ "Client" ],
        "summary": "Создание нового клиента",
        "operationId": "ClientCreate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное создание клиента",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "ID созданного клиента"
                }
              }
            }
          },
          "400": {
            "description": "Неверный запрос"
          },
          "500": {
            "description": "Внутренняя ошибка сервера"
          }
        }
      }
    },
    "/api/Client/update": {
      "put": {
        "tags": [ "Client" ],
        "summary": "Обновление данных клиента",
        "operationId": "ClientUpdate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClientRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Успешное обновление",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "Количество затронутых записей"
                }
              }
            }
          },
          "404": {
            "description": "Клиент не найден"
          }
        }
      }
    },
    "/api/Client/delete": {
      "delete": {
        "tags": [ "Client" ],
        "summary": "Удаление клиента",
        "operationId": "ClientDelete",
        "parameters": [
          {
            "name": "clientId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "ID клиента"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешное удаление",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "404": {
            "description": "Клиент не найден"
          }
        }
      }
    },
...
    "/api/Client/get/{clientId}": {
      "get": {
        "tags": [ "Client" ],
        "summary": "Получение клиента по ID",
        "operationId": "ClientGetById",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "description": "ID клиента"
          }
        ],
        "responses": {
          "200": {
            "description": "Успешный запрос",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClientResponse"
                }
              }
            }
          },
          "404": {
            "description": "Клиент не найден"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateClientRequest": {
        "type": "object",
        "properties": {
          "document": {
            "type": "string",
            "description": "Паспортные данные"
          },
          "surName": {
            "type": "string",
            "description": "Фамилия"
          },
          "firstName": {
            "type": "string",
            "description": "Имя"
          },
          "patronymic": {
            "type": "string",
            "nullable": true,
            "description": "Отчество"
          },
          "birthday": {
            "type": "string",
            "format": "date-time",
            "description": "Дата рождения"
          }
        },
        "required": [ "document", "surName", "firstName", "birthday" ]
      },
      "UpdateClientRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32",
            "description": "ID клиента"
          },
          "document": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [ "clientId", "document", "surName", "firstName", "birthday" ]
      },
      "ClientResponse": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "integer",
            "format": "int32"
          },
          "document": {
            "type": "string"
          },
          "surName": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "patronymic": {
            "type": "string",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}